Compute distribution of FvFm and PSII.

Organises the raw data output for 1 or more plates and plots the distribution.

Usage

computeFvFm_PSII(
  data,
  plate_names,
  layout,
  smartsheet,
  copynumber,
  output_location,
  width = 20,
  height = 21
)

Arguments

data

data.frame or list of data.frame; from FluorCam software (frames-numeric)

plate_names

character; name of plates.

layout

data.frame or list of data.frame; plate layout.

smartsheet

data.frame; selection results, ie BASTA.

copynumber

data.frame; copy number results.

output_location

path; location to save output files

width

numeric; pdf width size. deafult is 20.

height

numeric;pdf height size. deafult is 21.

Value

Returns a list of dataframes, where each plates has a dataframe for FvFm and PSII. Saves an excel file containing all data and a plot for FvFm and PSII.

Details

Takes as input a single dataframe or a list of dataframe. Save data as chlorofluoro_dataset_1.xlsx and plots as chlorofluoro_plot_1.pdf in desired directory.

Examples

data("CF_demodata")
data("layout")
data("selection_results")
data("copynumber")

output_location <-  tempdir()
step_one <- computeFvFm_PSII(CF_demodata,
plate_names = names(CF_demodata),
layout,
smartsheet = selection_results,
copynumber,
output_location)
#> Data saved to /tmp/Rtmpmm6ZbD/chlorofluoro_dataset_1.xlsx
#> Plots saved to /tmp/Rtmpmm6ZbD/chlorofluoro_plot_1.pdf
#> Returned data